+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## WIP
+### Fixed
+- Handle list form in query properly in command line shell [#42]
+
 ## 0.4.4
 ### Changed
 - [Breaking] Consonidated all user facing functions to `datalevin.core`, so users don't have to understand and require different namespaces in order to use all features.